# 12_Backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ScottsTechX Commerce OS API",
    "version": "1.0.0",
    "description": "Trust-gated marketplace backend for the Uganda MVP slice.\n\nThis spec is the source of truth for clients (Flutter, web, admin). It mirrors\nthe implementation in `12_Backend/src/server.ts` and the migration-derived\nschema. Endpoints marked `STUB` return HTTP 501 with `X-Stub-Reason` header\nand do NOT perform any external I/O.\n\n## Authentication\n\nAll endpoints require `Authorization: Bearer ***` where the token is an\nHS256 JWT with the following claims:\n  - `iss`: must equal `scottstechx`\n  - `aud`: must equal `scottstechx-api`\n  - `sub`: user's UUID\n  - `role`: one of `buyer | driver | seller | admin`\n  - `exp`: present and not in the past\n\n## Idempotency\n\n`POST` and `PATCH` endpoints that change state require an `Idempotency-Key`\nheader of at least 8 characters. Replays with the same body return the\noriginal response. Replays with a different body under the same key return\nHTTP 409.\n\n## Error format\n\nAll errors return JSON: `{ error: string, message: string, details?: any }`.\nThe `error` field is a stable machine-readable code.",
    "license": { "name": "Proprietary" }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local dev" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["phone", "password"],
        "properties": {
          "phone": { "type": "string", "minLength": 6, "maxLength": 30 },
          "password": { "type": "string", "maxLength": 200 },
          "role": { "type": "string", "description": "Accepted but ignored; server is authoritative" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["token", "userId", "role", "trustTier", "expiresAt"],
        "properties": {
          "token": { "type": "string" },
          "userId": { "type": "string", "format": "uuid" },
          "role": { "type": "string", "enum": ["buyer", "driver", "seller", "admin"] },
          "trustTier": { "type": "string", "enum": ["BRONZE", "SILVER", "GOLD", "PLATINUM"] },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "FirebaseAuthRequest": {
        "type": "object",
        "required": ["idToken"],
        "properties": {
          "idToken": { "type": "string", "minLength": 20 },
          "role": { "type": "string", "enum": ["buyer", "driver", "seller"] }
        }
      },
      "FirebaseAuthResponse": {
        "type": "object",
        "required": ["token", "userId", "role", "email", "trustTier", "expiresAt"],
        "properties": {
          "token": { "type": "string" },
          "userId": { "type": "string", "format": "uuid" },
          "role": { "type": "string", "enum": ["buyer", "driver", "seller", "admin"] },
          "email": { "type": "string", "nullable": true },
          "trustTier": { "type": "string", "enum": ["BRONZE", "SILVER", "GOLD", "PLATINUM"] },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"], "summary": "Phone + password login, returns a JWT", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
        "responses": {
          "200": { "description": "JWT issued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } } },
          "400": { "description": "Validation error" },
          "401": { "description": "Invalid phone or password" }
        }
      }
    },
...
    "/api/v1/auth/firebase": {
      "post": {
        "tags": ["auth"], "summary": "Verify a Firebase Auth ID token, return our JWT", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FirebaseAuthRequest" } } } },
        "responses": {
          "200": { "description": "JWT issued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FirebaseAuthResponse" } } } },
          "400": { "description": "Validation error" },
          "401": { "description": "Invalid or expired Firebase token" },
          "503": { "description": "Firebase Auth not configured" }
        }
      }
    },
...
  }
}